moved models to root so that cloud functions and app can share models
