+++ functions/package.json
@@ -12,7 +12,7 @@
   "engines": {
     "node": "22"
   },
-  "main": "lib/index.js",
+  "main": "lib/functions/src/index.js",
   "dependencies": {
     "firebase-admin": "^12.7.0",
     "firebase-functions": "^6.0.1"
