+++ models/attendedEventsType.ts
