+++ models/beneficiaryType.ts
@@ -1,6 +1,6 @@
 import type { Timestamp } from "firebase/firestore";
-import type { AttendedEvents } from "./attendedEventsType";
-import type { Guardian } from "./guardianType";
+import type { AttendedEvents } from "./attendedEventsType.js";
+import type { Guardian } from "./guardianType.js";
 
 export interface Beneficiary {
   docID: string;
