+++ vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vite'
 import tailwindcss from '@tailwindcss/vite'
 import react from '@vitejs/plugin-react'
 import { qrcode } from 'vite-plugin-qrcode';
+import path from 'path';
 
 
 // https://vite.dev/config/
@@ -11,4 +12,9 @@ export default defineConfig({
     tailwindcss(),
     qrcode() // only applies in dev mode
   ],
+  resolve: {
+    alias: {
+      '@models': path.resolve(__dirname, '../models'),
+    }
+  }
 })
